create design component
